+++ src/styles.css
@@ -2,7 +2,7 @@
 @import url("sidebar/css/sidebar.css");
 @import url("oracles/css/oracles.css");
 @import url("tracks/css/tracks.css");
-@import url("tracks/css/clocks.css");
+@import url("clocks/css/clocks.css");
 
 .theme-light {
   --color-code: 0, 0, 0;
